deployment_scenarios.xml: adds rmt ssl folder volume
The registry conf.yml file points to certificate under `/etc/rmt/ssl`
folder. However, the command to run the registry does not create a
volume to that directory in the registry container. Thus, the container
does not have access to the certificate. This commit adds that volume to
allow the registry read the necessary file. Close #282

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>