Update the links to the SLE15SP1 media download
SLE15SP1 is finally released. The page linked is now showing the beta of
SLE12 SP5, which is definitely not we want our users to install.

The link has been updated to the JeOS page, from there the user has to
click either the "how to buy" or the "60 day free trial" buttons. We
cannot embed the resulting URLs because they are generated on the fly
and they will definitely break over the time.

The asciidoc variable holding the URL of the media has been renamed to
be more specific: this is about JeOS which is relevant only for certain
types of target deployments of CaaSP (VMware, OpenStack). For bare metal
deployments the link will be a different one, hence the asciidoc will
have to use a different variable to hold the download URL.

The wording has also been changed inside of the VMware and OpenStack
deployment to make it consistent.