Network table: drop worker node section
Both dex-oidc and gangway services are reachable on all nodes of the
cluster. Move their requirement to the "all nodes" section.

As a result of this action there are no special rules for the worker
nodes. Basically all the nodes of the cluster have the same set of basic
rules. The control plane has some special ones in addition to them.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>