Add information about loadbalancer and master nodes
When we are adding a master node that was not originally planned
to be added from the beginning of the cluster (e.g. buy new 10
servers) they will **NOT** be automatically added to the config
of the loadbalancer.

The same happens when you delete a master node. There will be
stale entries pointing to a non-cluster resource unless the use
manually removes them.