Update deployment guide to support KVM
The existing bare metal deployment guide is also sufficient for
deployment on KVM machines. This change adds KVM to the list of
supported platforms, makes minor modifications to the deployment
documentation to make that support explicit, adds a note about how to
manage virtual machines, and fixes the load balancer instructions to be
clear that the guide on setting up a load balancer is applicable to both
bare metal and KVM hosts.