Warning about multiples kubeconfig with the same user
Recently we have a customer trying to use multiples kubeconfig file with
the same user. This cause issue because after one of the machines with
the files refresh the OIDC token, the rest cannot refresh the token
anymore. This commit adds a warning about this situation.

Issue https://github.com/SUSE/doc-caasp/issues/735

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>