Update admission controller info in GPU guide
In v5 skuba will enable the ExtendedResourceToleration admission
controller by default, so users will not have to manually enable it and
maintain it through upgrades. This patch removes these now-unnecessary
instructions, makes the discussion of the admission controller more
prominent, and adds further clarification on the effects of tainting
nodes.