adoc: clarify patterns-caasp-Management installation node
We should clearly state that the `patterns-caasp-Management` package is
installed on the management node, and not on master/worker nodes. This is
important because we have removed the patterns for those nodes (so it might be
confusing during the upgrade).

The exact wording has been taken from `adoc/deployment-aws.adoc`, and by copying
that we become a bit more consistent.

See SUSE/avant-garde#1878

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>